demo-apps/ml-tools/requirements.txt
streamlit
numpy
tensorflow
Pillow
scikit-learn
